test with zero-ing out garbage
